requirements.txt
google-adk
crawl4ai
aiosqlite
sqlalchemy
nest_asyncio
python-dotenv
certifi
litellm

